+++ dlls/monster/CHGrunt.cpp
@@ -290,7 +290,7 @@ void CHGrunt::IdleSound(void)
 }
 
 void CHGrunt::PlaySentenceSound(int sentenceType) {
-	if (sentenceType >= ARRAYSIZE(pGruntSentences)) {
+	if (sentenceType >= (int)ARRAYSIZE(pGruntSentences)) {
 		return;
 	}
 	SENTENCEG_PlayRndSz(ENT(pev), pGruntSentences[sentenceType], SENTENCE_VOLUME, GRUNT_ATTN, 0, m_voicePitch);
